pyproject.toml
[project]
name = "secure-ai-agent-matching-platform"
version = "0.1.0"
description = "Secure AI Agent Matching Platform with A2A Protocol"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]==1.6.1",
    "a2a-sdk==0.3.4",
    "httpx",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
